+++ tests/types/types.go
@@ -94,18 +94,20 @@ type MntDevice struct {
 }
 
 type Test struct {
-	Name              string
-	In                []Disk // Disk state before running Ignition
-	Out               []Disk // Expected disk state after running Ignition
-	MntDevices        []MntDevice
-	SystemDirFiles    []File
-	Env               []string // Environment variables for Ignition
-	Config            string
-	ConfigMaxVersion  string
-	ConfigMinVersion  string
-	ConfigVersion     string
-	ConfigShouldBeBad bool // Set to true to skip config validation step
-	SkipCriticalCheck bool // Set to true to skip critical logging check
+	Name                  string
+	In                    []Disk // Disk state before running Ignition
+	Out                   []Disk // Expected disk state after running Ignition
+	MntDevices            []MntDevice
+	SystemDirFiles        []File
+	SystemRuntimeDirFiles []File
+	SystemLocalDirFiles   []File
+	Env                   []string // Environment variables for Ignition
+	Config                string
+	ConfigMaxVersion      string
+	ConfigMinVersion      string
+	ConfigVersion         string
+	ConfigShouldBeBad     bool // Set to true to skip config validation step
+	SkipCriticalCheck     bool // Set to true to skip critical logging check
 }
 
 func (ps Partitions) GetPartition(label string) *Partition {
@@ -295,6 +297,14 @@ func DeepCopy(t Test) Test {
 	SystemDirFiles := make([]File, len(t.SystemDirFiles))
 	copy(SystemDirFiles, t.SystemDirFiles)
 	t.SystemDirFiles = SystemDirFiles
+
+	SystemRuntimeDirFiles := make([]File, len(t.SystemRuntimeDirFiles))
+	copy(SystemRuntimeDirFiles, t.SystemRuntimeDirFiles)
+	t.SystemRuntimeDirFiles = SystemRuntimeDirFiles
+
+	SystemLocalDirFiles := make([]File, len(t.SystemLocalDirFiles))
+	copy(SystemLocalDirFiles, t.SystemLocalDirFiles)
+	t.SystemLocalDirFiles = SystemLocalDirFiles
 
 	return t
 }